+++ merlin/study/celeryadapter.py
@@ -6,7 +6,7 @@
 #
 # LLNL-CODE-797170
 # All rights reserved.
-# This file is part of Merlin, Version: 1.11.0.
+# This file is part of Merlin, Version: 1.11.1.
 #
 # For details, see https://github.com/LLNL/merlin.
 #
@@ -37,14 +37,21 @@ import socket
 import subprocess
 import time
 from contextlib import suppress
+from typing import Dict, List, Optional
 
 from tabulate import tabulate
 
+from amqp.exceptions import ChannelError
+from celery import Celery
+
+from merlin.config import Config
 from merlin.study.batch import batch_check_parallel, batch_worker_launch
 from merlin.utils import apply_list_of_regex, check_machines, get_procs, get_yaml_var, is_running
 
 
 LOG = logging.getLogger(__name__)
+
+# TODO figure out a better way to handle the import of celery app and CONFIG
 
 
 def run_celery(study, run_mode=None):
@@ -71,23 +78,31 @@ def run_celery(study, run_mode=None):
     queue_merlin_study(study, adapter_config)
 
 
-def get_running_queues():
+def get_running_queues(celery_app_name: str, test_mode: bool = False) -> List[str]:
     """
-    Check for running celery workers with -Q queues
-    and return a unique list of the queues
+    Check for running celery workers by looking at the currently running processes.
+    If there are running celery workers, we'll pull the queues from the -Q tag in the
+    process command. The list returned here will contain only unique celery queue names.
+    This must be run on the allocation where the workers are running.
 
-    Must be run on the allocation where the workers are running
+    :param `celery_app_name`: The name of the celery app (typically merlin here unless testing)
+    :param `test_mode`: If True, run this function in test mode
+    :returns: A unique list of celery queues with workers attached to them
     """
     running_queues = []
 
-    if not is_running("celery worker"):
+    if not is_running(f"{celery_app_name} worker"):
         return running_queues
 
-    procs = get_procs("celery")
+    proc_name = "celery" if not test_mode else "sh"
+    procs = get_procs(proc_name)
     for _, lcmd in procs:
         lcmd = list(filter(None, lcmd))
         cmdline = " ".join(lcmd)
         if "-Q" in cmdline:
+            if test_mode:
+                echo_cmd = lcmd.pop(2)
+                lcmd.extend(echo_cmd.split())
             running_queues.extend(lcmd[lcmd.index("-Q") + 1].split(","))
 
     running_queues = list(set(running_queues))
@@ -95,7 +110,7 @@ def get_running_queues():
     return running_queues
 
 
-def get_queues(app):
+def get_active_celery_queues(app):
     """Get all active queues and workers for a celery application.
 
     Unlike get_running_queues, this goes through the application's server.
@@ -110,7 +125,7 @@ def get_queues(app):
     :example:
 
     >>> from merlin.celery import app
-    >>> queues, workers = get_queues(app)
+    >>> queues, workers = get_active_celery_queues(app)
     >>> queue_names = [*queues]
     >>> workers_on_q0 = queues[queue_names[0]]
     >>> workers_not_on_q0 = [worker for worker in workers
@@ -132,7 +147,7 @@ def get_queues(app):
 
 def get_active_workers(app):
     """
-    This is the inverse of get_queues() defined above. This function
+    This is the inverse of get_active_celery_queues() defined above. This function
     builds a dict where the keys are worker names and the values are lists
     of queues attached to the worker.
 
@@ -157,19 +172,20 @@ def get_active_workers(app):
     return worker_queue_map
 
 
-def celerize_queues(queues):
+def celerize_queues(queues: List[str], config: Optional[Dict] = None):
     """
     Celery requires a queue tag to be prepended to their
     queues so this function will 'celerize' every queue in
     a list you provide it by prepending the queue tag.
 
-    :param `queues`: A list of queues that need the queue
-                     tag prepended.
+    :param `queues`: A list of queues that need the queue tag prepended.
+    :param `config`: A dict of configuration settings
     """
-    from merlin.config.configfile import CONFIG  # pylint: disable=C0415
+    if config is None:
+        from merlin.config.configfile import CONFIG as config  # pylint: disable=C0415
 
     for i, queue in enumerate(queues):
-        queues[i] = f"{CONFIG.celery.queue_tag}{queue}"
+        queues[i] = f"{config.celery.queue_tag}{queue}"
 
 
 def _build_output_table(worker_list, output_table):
@@ -219,7 +235,7 @@ def query_celery_workers(spec_worker_names, queues, workers_regex):
     # --queues flag
     if queues:
         # Get a mapping between queues and the workers watching them
-        queue_worker_map, _ = get_queues(app)
+        queue_worker_map, _ = get_active_celery_queues(app)
         # Remove duplicates and prepend the celery queue tag to all queues
         queues = list(set(queues))
         celerize_queues(queues)
@@ -261,26 +277,54 @@ def query_celery_workers(spec_worker_names, queues, workers_regex):
     print()
 
 
-def query_celery_queues(queues):
-    """Return stats for queues specified.
-
-    Send results to the log.
+def query_celery_queues(queues: List[str], app: Celery = None, config: Config = None) -> Dict[str, List[str]]:
     """
-    from merlin.celery import app  # pylint: disable=C0415
+    Build a dict of information about the number of jobs and consumers attached
+    to specific queues that we want information on.
 
-    connection = app.connection()
-    found_queues = []
-    try:
-        channel = connection.channel()
-        for queue in queues:
-            try:
-                name, jobs, consumers = channel.queue_declare(queue=queue, passive=True)
-                found_queues.append((name, jobs, consumers))
-            except Exception as e:  # pylint: disable=C0103,W0718
-                LOG.warning(f"Cannot find queue {queue} on server.{e}")
-    finally:
-        connection.close()
-    return found_queues
+    :param queues: A list of the queues we want to know about
+    :param app: The celery application (this will be none unless testing)
+    :param config: The configuration object that has the broker name (this will be none unless testing)
+    :returns: A dict of info on the number of jobs and consumers for each queue in `queues`
+    """
+    if app is None:
+        from merlin.celery import app  # pylint: disable=C0415
+    if config is None:
+        from merlin.config.configfile import CONFIG as config  # pylint: disable=C0415
+
+    # Initialize the dictionary with the info we want about our queues
+    queue_info = {queue: {"consumers": 0, "jobs": 0} for queue in queues}
+
+    # Open a connection via our Celery app
+    with app.connection() as conn:
+        # Open a channel inside our connection
+        with conn.channel() as channel:
+            # Loop through all the queues we're searching for
+            for queue in queues:
+                try:
+                    # Count the number of jobs and consumers for each queue
+                    _, queue_info[queue]["jobs"], queue_info[queue]["consumers"] = channel.queue_declare(
+                        queue=queue, passive=True
+                    )
+                # Redis likes to throw this error when a queue we're looking for has no jobs
+                except ChannelError:
+                    pass
+
+    # Redis doesn't keep track of consumers attached to queues like rabbit does
+    # so we have to count this ourselves here
+    if config.broker.name in ("rediss", "redis"):
+        # Get a dict of active queues by querying the celery app
+        active_queues = app.control.inspect().active_queues()
+        if active_queues is not None:
+            # Loop through each active queue that was found
+            for active_queue_list in active_queues.values():
+                # Loop through each queue that each worker is watching
+                for active_queue in active_queue_list:
+                    # If this is a queue we're looking for, increment the consumer count
+                    if active_queue["name"] in queues:
+                        queue_info[active_queue["name"]]["consumers"] += 1
+
+    return queue_info
 
 
 def get_workers_from_app():
@@ -297,6 +341,27 @@ def get_workers_from_app():
     if workers is None:
         return []
     return [*workers]
+
+
+def check_celery_workers_processing(queues_in_spec: List[str], app: Celery) -> bool:
+    """
+    Query celery to see if any workers are still processing tasks.
+
+    :param queues_in_spec: A list of queues to check if tasks are still active in
+    :param app: The celery app that we're querying
+    :returns: True if workers are still processing tasks, False otherwise
+    """
+    # Query celery for active tasks
+    active_tasks = app.control.inspect().active()
+
+    # Search for the queues we provided if necessary
+    if active_tasks is not None:
+        for tasks in active_tasks.values():
+            for task in tasks:
+                if task["delivery_info"]["routing_key"] in queues_in_spec:
+                    return True
+
+    return False
 
 
 def _get_workers_to_start(spec, steps):
@@ -463,7 +528,7 @@ def start_celery_workers(spec, steps, celery_args, disable_logs, just_return_com
         running_queues.extend(local_queues)
         queues = queues.split(",")
         if not overlap:
-            running_queues.extend(get_running_queues())
+            running_queues.extend(get_running_queues("merlin"))
             # Cache the queues from this worker to use to test
             # for existing queues in any subsequent workers.
             # If overlap is True, then do not check the local queues.
@@ -611,7 +676,7 @@ def stop_celery_workers(queues=None, spec_worker_names=None, worker_regex=None):
     from merlin.celery import app  # pylint: disable=C0415
 
     LOG.debug(f"Sending stop to queues: {queues}, worker_regex: {worker_regex}, spec_worker_names: {spec_worker_names}")
-    active_queues, _ = get_queues(app)
+    active_queues, _ = get_active_celery_queues(app)
 
     # If not specified, get all the queues
     if queues is None: